Refactoring: made a constants folder where defaults and MixerProtocolPresets.js are placed. Added default mixerProtocol if an unknown protocol (e.g. typo in settings) is selected. Fixed last iteration in FadeInOut, it always did one extra iteration after target value was met.
